Clean-up follow-up: Move dynamic elements inside its own package before too much stuff are put everywhere
